chore: removes date picker question code from sideloading
