+++ packages/surveys/src/lib/styles.ts
@@ -1,17 +1,21 @@
 import global from "@/styles/global.css?inline";
 import preflight from "@/styles/preflight.css?inline";
+import calendarCss from "react-calendar/dist/Calendar.css?inline";
+import datePickerCss from "react-date-picker/dist/DatePicker.css?inline";
 
 import { isLight, mixColor } from "@formbricks/lib/utils";
 import { TProductStyling } from "@formbricks/types/product";
 import { TSurveyStyling } from "@formbricks/types/surveys";
 
 import editorCss from "../../../ui/Editor/stylesEditorFrontend.css?inline";
+import datePickerCustomCss from "../styles/date-picker.css?inline";
 
 export const addStylesToDom = () => {
   if (document.getElementById("formbricks__css") === null) {
     const styleElement = document.createElement("style");
     styleElement.id = "formbricks__css";
-    styleElement.innerHTML = preflight + global + editorCss;
+    styleElement.innerHTML =
+      preflight + global + editorCss + datePickerCss + calendarCss + datePickerCustomCss;
     document.head.appendChild(styleElement);
   }
 };
