+++ packages/surveys/src/styles/date-picker.css
